Fix handling of dest charsets with chars > u7FFF and missing chasets
Prevoisly, if unexisting charset was specified in the command line,
catdoc/xls2csv/catppt silently falls back to default charset. Now it
exits with non-zero exit code and appropriate message

It was also found that in make_reverse_map unicode code was incorrectly
casted to (unsigned) instead of (unsigned short) which cause segfault
if charater with 16-bit unicode code highest bit set present in the
destination charset (i.e. in mac-roman).

+++ src/charsets.c
@@ -51,7 +51,7 @@ CHARSET make_reverse_map(short int *charset) {
 	}	
 	for (i=0;i<256;i++) {
 		k= charset[i];
-		j=  (unsigned)k>>8;
+		j=  (unsigned short int)k>>8;
 		if (!newmap[j]) {
 			newmap[j] = (short int *)malloc(sizeof(short int)*256);
 			if (!newmap[j]) {
